navibuVE/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.3
Werkzeug==2.3.7 